Switch over to calling ResolveSecrets
