+++ pkg/process/config/config_test.go
@@ -327,25 +327,6 @@ func TestEnvSiteConfig(t *testing.T) {
 	os.Unsetenv("DD_PROCESS_AGENT_URL")
 }
 
-func TestDontOverwriteEnvVarsWithEncryptedSecrets(t *testing.T) {
-	assert := assert.New(t)
-	defer restoreGlobalConfig()
-
-	config.Datadog = config.NewConfig("datadog", "DD", strings.NewReplacer(".", "_"))
-	os.Setenv("DD_API_KEY", "ABC")
-	loadEnvVariables()
-	os.Unsetenv("DD_API_KEY")
-
-	assert.True(config.Datadog.IsSet("api_key"))
-
-	config.Datadog = config.NewConfig("datadog", "DD", strings.NewReplacer(".", "_"))
-	os.Setenv("DD_API_KEY", "ENC[ABC]")
-	loadEnvVariables()
-	os.Unsetenv("DD_API_KEY")
-
-	assert.False(config.Datadog.IsSet("api_key"))
-}
-
 func TestIsAffirmative(t *testing.T) {
 	value, err := isAffirmative("yes")
 	assert.Nil(t, err)
